split lexer into a separate module
